refactor : DefaultFarePolicy -> DistanceFarePolicy 로 이름 변경
